requirements.txt
pandas==2.3.1
plotly==6.2.0
python-dotenv==1.1.1
Requests==2.32.4
streamlit==1.28.2
tqdm==4.66.5
